requirements.txt
streamlit>=1.20.0
numpy>=1.21.0
torch>=1.10.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
wordcloud>=1.8.0
PyPDF2>=2.0.0
pdfplumber>=0.7.0
python-docx>=0.8.11
nltk>=3.6.0
spacy>=3.0.0
